+++ src/ini-parser.c
@@ -99,7 +99,7 @@ struct ini_file *read_slidedat_ini_file(const char *path, const char *ini_filena
 
             // parse entry
             struct ini_entry *temp_entry = (struct ini_entry *)malloc(sizeof(struct ini_entry));
-            if (!contains(buffer, "=")) {
+            if (strstr(buffer, "=") == NULL) {
                 continue;
             }
             char **splitted_entry = str_split(buffer, '=');