+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.4.5
 
-* bugfix for string contains function in utils
+* bugfix for string contains function in utils (replaced by string.h's `strstr` function)
 
 ## 0.4.4